// public/CroncatTasks.types.ts
export type CosmosMsgForEmpty = {
  bank: BankMsg;
} | {
  custom: Empty;
} | {
  staking: StakingMsg;
} | {
  distribution: DistributionMsg;
} | {
  stargate: {
    type_url: string;
    value: Binary;
    [k: string]: unknown;
  };
} | {
  ibc: IbcMsg;
} | {
  wasm: WasmMsg;
} | {
  gov: GovMsg;
};
export type BankMsg = {
  send: {
    amount: Coin[];
    to_address: string;
    [k: string]: unknown;
  };
} | {
  burn: {
    amount: Coin[];
    [k: string]: unknown;
  };
};
export type Uint128 = string;
// ...
export type Boundary = {
  height: BoundaryHeight;
} | {
  time: BoundaryTime;
};
export type Interval = "once" | "immediate" | {
  block: number;
} | {
  cron: string;
};
export type CosmosQueryForWasmQuery = {
  croncat: CroncatQuery;
} | {
  wasm: WasmQuery;
};
export type WasmQuery = {
  smart: {
    contract_addr: string;
    msg: Binary;
    [k: string]: unknown;
  };
} | {
  raw: {
    contract_addr: string;
    key: Binary;
    [k: string]: unknown;
  };
} | {
  contract_info: {
    contract_addr: string;
    [k: string]: unknown;
  };
};
export type ValueIndex = {
  key: string;
} | {
  index: number;
};
export type PathToValue = ValueIndex[];
export interface UpdateConfigMsg {
  croncat_agents_key?: [string, [number, number]] | null;
  croncat_factory_addr?: string | null;
  croncat_manager_key?: [string, [number, number]] | null;
  gas_action_fee?: number | null;
  gas_base_fee?: number | null;
  gas_limit?: number | null;
  gas_query_fee?: number | null;
  slot_granularity_time?: number | null;
}
export interface TaskRequest {
  actions: ActionForEmpty[];
  boundary?: Boundary | null;
  cw20?: Cw20Coin | null;
  interval: Interval;
  queries?: CosmosQueryForWasmQuery[] | null;
  stop_on_fail: boolean;
  transforms?: Transform[] | null;
}
export interface ActionForEmpty {
  gas_limit?: number | null;
  msg: CosmosMsgForEmpty;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export interface Empty {
  [k: string]: unknown;
}
export interface IbcTimeout {
  block?: IbcTimeoutBlock | null;
  timestamp?: Timestamp | null;
  [k: string]: unknown;
}
export interface IbcTimeoutBlock {
  height: number;
  revision: number;
  [k: string]: unknown;
}
export interface BoundaryHeight {
  end?: Uint64 | null;
  start?: Uint64 | null;
}
export interface BoundaryTime {
  end?: Timestamp | null;
  start?: Timestamp | null;
}
export interface Cw20Coin {
  address: string;
  amount: Uint128;
}
export interface CroncatQuery {
  check_result: boolean;
  contract_addr: string;
  msg: Binary;
}
export interface Transform {
  action_idx: number;
  action_path: PathToValue;
  query_idx: number;
  query_response_path: PathToValue;
}